requirements.txt
numpy
Cython
python-telegram-bot
python-opencv
moviepy
insightface
onnx
onnxruntime
